+++ src/stubs/github/github.ts
@@ -1,6 +1,6 @@
 /**
- * Last Reviewed Commit: https://github.com/actions/toolkit/blob/f58042f9cc16bcaa87afaa86c2974a8c771ce1ea/packages/github/src/github.ts
- * Last Reviewed Date: 2025-09-10
+ * Last Reviewed Commit: 94de2cf6d446e4afd3d7b66408e7c513633dddaa
+ * Last Reviewed Date: 2026-01-16
  *
  * @remarks
  *